fix(dataset): addressing pr comments

+++ include/vision/dataset.hpp
@@ -37,13 +37,13 @@ public:
     // Static methos
     static std::vector<std::string> ListAllStoredDatasets();
     static bool isValidConfigurationFile(const std::string& config_file_path);
-    static bool isValidConfiguration(double max_period, uint16_t nb_frames, CAPTURE_MODE capture_mode, IMU_COLLECTION_MODE imu_collection_mode,
+    static bool isValidConfiguration(double max_period, uint8_t nb_frames, CAPTURE_MODE capture_mode, IMU_COLLECTION_MODE imu_collection_mode,
                                     uint8_t image_capture_rate, float imu_sample_rate_hz, ProcessingStage target_processing_stage,
                                     uint64_t capture_start_time);
 
     // Getters
     uint64_t GetCaptureStartTime() const { return capture_start_time; }
-    uint16_t GetTargetFrameNb() const { return target_frame_nb; }
+    uint8_t GetTargetFrameNb() const { return target_frame_nb; }
     double GetMaximumPeriod() const { return maximum_period; }
     CAPTURE_MODE GetDatasetCaptureMode() const { return dataset_capture_mode; }
     IMU_COLLECTION_MODE GetIMUCollectionMode() const { return imu_collection_mode; }
@@ -62,7 +62,7 @@ public:
     bool OverlapsWith(const Dataset& other) const;
 
     // Actual constructors ~ not to be used
-    Dataset(double max_period, uint16_t nb_frames, CAPTURE_MODE capture_mode, IMU_COLLECTION_MODE imu_collection_mode,
+    Dataset(double max_period, uint8_t nb_frames, CAPTURE_MODE capture_mode, IMU_COLLECTION_MODE imu_collection_mode,
             uint8_t image_capture_rate, float imu_sample_rate_hz, ProcessingStage target_processing_stage,
             uint64_t capture_start_time);
 
@@ -77,7 +77,7 @@ private:
     
     uint64_t capture_start_time; // unix in ms. For scheduling
     double maximum_period;
-    uint16_t target_frame_nb;
+    uint8_t target_frame_nb;
     CAPTURE_MODE dataset_capture_mode;
     IMU_COLLECTION_MODE imu_collection_mode;
     uint8_t image_capture_rate; // [s]
